Update baseresponse.go
Hello, when I use elastigo found inside the package is the core of search methods. When the result of elasticsearch returned and the reason is that this field appears when the field is not a string type, will json Unmarshal error occurs, I will reason the string replaced json.RawMessage, solves this problem.